Phase 1: Fix critical correctness issues
- Fix weight loading API: signature now requires config parameter
- Remove silent fallback to random weights in LLMEngine
- Fix paged attention slot_mapping (use compute_slot_mapping)
- Fix WINDOW_SIZE bug (attend to ALL positions, not just 0-127)
- Fix Metal compatibility (use float32 default, manual gather)
- Add test for non-identity block tables
- Add test for long sequences (>128 tokens)
- Update documentation with softened language
- Update PLAN.md with accurate status

Tests: KV cache 7/7, Layer parity 6/6, MTP structural tests pass

+++ nanovllm_jax/config.py
@@ -52,8 +52,8 @@ class Qwen3_5Config:
     attention_bias: bool = False
     tie_word_embeddings: bool = True
     
-    # Computation dtype (bfloat16 for CPU/CUDA, float16 for Metal)
-    dtype: str = "bfloat16"  # Options: "bfloat16", "float16", "float32"
+    # Computation dtype (bfloat16 for CPU/CUDA, float32 for Metal)
+    dtype: str = "float32"  # Options: "bfloat16", "float16", "float32"
     
     # MTP config
     mtp_num_hidden_layers: int = 1
